remove 23KiBs of installer bloat
NSIS has a "new" look called MUI which is recommended by the docs.
However, the old one is faster and produces a 23KiB smaller executable
as of right now. Clearly the old look is better than the new look.

Also makes both installers artifacts in the CI and adds the subprojects
to .gitignore.